+++ lib/importer.dart
@@ -6,5 +6,7 @@ export 'package:sbj_coffee/coffee_details/blond_roast.dart';
 export 'package:sbj_coffee/coffee_details/dark_roast.dart';
 export 'package:sbj_coffee/coffee_details/medium_roast.dart';
 
+export 'package:sbj_coffee/bottom_bar.dart';
+
 export 'package:flutter/material.dart';
 export 'package:curved_navigation_bar/curved_navigation_bar.dart';
